Moved HeroChat to its own logging class.
Added HeroChat as a dependency.
Added code to register HeroChat event listening if HeroChat is present.
