+++ src/de/diddiz/LogBlock/listeners/ChatLogging.java
@@ -1,6 +1,5 @@
 package de.diddiz.LogBlock.listeners;
 
-import com.dthielke.herochat.ChannelChatEvent;
 import static de.diddiz.LogBlock.config.Config.isLogging;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
@@ -21,13 +20,6 @@ public class ChatLogging extends LoggingListener
 		if (isLogging(event.getPlayer().getWorld(), Logging.CHAT))
 			consumer.queueChat(event.getPlayer().getName(), event.getMessage());
 	}
-        
-        @EventHandler(priority = EventPriority.MONITOR)
-        public void onChannelChatEvent(ChannelChatEvent event) {
-            PlayerChatEvent bukkitEvent = event.getBukkitEvent();
-            if (isLogging(bukkitEvent.getPlayer().getWorld(), Logging.CHAT))
-			consumer.queueChat(bukkitEvent.getPlayer().getName(), bukkitEvent.getMessage());
-        }
 
 	@EventHandler(priority = EventPriority.MONITOR)
 	public void onPlayerChat(PlayerChatEvent event) {
